refactor: split user API module by responsibility
